requirements.txt
fastapi==0.115.0
uvicorn==0.30.0
fastmcp==0.4.1
google-generativeai==0.8.4
neo4j==5.17.0
PyGithub==2.3.0
python-dotenv==1.0.1
httpx==0.27.0
pydantic==2.7.4
websockets==12.0
